+++ src/app/controller/collection.ts
@@ -1,8 +1,8 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import { BadRequest, Conflict, NotFound, UnprocessableContent } from '@curveball/http-errors';
+import { BadRequest, Conflict, UnprocessableContent } from '@curveball/http-errors';
 import * as hal from '../formats/hal.js';
-import { PrincipalService } from '../../principal/service.js';
+import * as services from '../../services.js';
 
 type NewPrincipalBody = {
   nickname: string;
@@ -10,11 +10,11 @@ type NewPrincipalBody = {
   type: 'user' | 'app' | 'group';
 }
 
-class UserCollectionController extends Controller {
+class AppCollectionController extends Controller {
 
   async get(ctx: Context) {
 
-    const principalService = new PrincipalService(ctx.privileges);
+    const principalService = new services.principal.PrincipalService(ctx.privileges);
     const apps = await principalService.findAll('app');
     ctx.response.body = hal.collection(apps);
 
@@ -22,32 +22,28 @@ class UserCollectionController extends Controller {
 
   async post(ctx: Context) {
 
-    const principalService = new PrincipalService(ctx.privileges);
+    const principalService = new services.principal.PrincipalService(ctx.privileges);
 
     ctx.request.validate<NewPrincipalBody>(
-      'https://curveballjs.org/schemas/a12nserver/new-principal.json'
+      'https://curveballjs.org/schemas/a12nserver/principal-new.json'
     );
 
     if (ctx.request.body.type !== 'app') {
       throw new BadRequest('You may only create principals with type "app" at this endpoint');
     }
 
-    const identity = ctx.request.links.get('me')?.href;
-    if (!identity) {
-      throw new UnprocessableContent('You must specify a link with rel "me", either via a HAL link or HTTP Link header');
-    }
-
-    try {
-      await principalService.findByIdentity(identity);
-      throw new Conflict('Principal already exists');
-    } catch (err) {
-      if (!(err instanceof NotFound)) {
-        throw err;
+    let identity = null;
+    if (ctx.request.links.has('me')) {
+      identity = ctx.request.links.get('me')!;
+      if (!identity.href.match(/^https?:(.*)$/)) {
+        throw new UnprocessableContent('App "me" URI must be a http or https URI');
+      }
+      if (await services.principalIdentity.findByUri(identity.href)) {
+        throw new Conflict(`The uri "${identity.href}" already exists`);
       }
     }
 
     const app = await principalService.save({
-      identity,
       nickname: ctx.request.body.nickname,
       type: 'app',
       active: ctx.request.body.active,
@@ -55,10 +51,21 @@ class UserCollectionController extends Controller {
       modifiedAt: new Date(),
     });
 
+    if (identity) {
+      await services.principalIdentity.create({
+        uri: identity.href,
+        principalId: app.id,
+        isPrimary: true,
+        label: identity.title ?? null,
+        markVerified: false,
+      });
+
+    }
+
     ctx.response.status = 201;
     ctx.response.headers.set('Location', app.href);
   }
 
 }
 
-export default new UserCollectionController();
+export default new AppCollectionController();