+++ src/principal-identity/controller/collection.ts
@@ -1,31 +1,28 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
 import { Forbidden } from '@curveball/http-errors';
-
 import * as hal from '../formats/hal.js';
-import * as userAppPermissionService from '../service.js';
-import { PrincipalService } from '../../principal/service.js';
+import * as services from '../../services.js';
 
-
-class UserAppPermissionsCollection extends Controller {
+class PrincipalIdentityCollection extends Controller {
 
   async get(ctx: Context) {
 
-    const principalService = new PrincipalService(ctx.privileges);
+    const principalService = new services.principal.PrincipalService(ctx.privileges);
     const principal = await principalService.findByExternalId(ctx.params.id, 'user');
+    const identities = await services.principalIdentity.findByPrincipal(principal);
 
     if (ctx.auth.equals(principal) && !ctx.privileges.has('admin')) {
       throw new Forbidden('You can only use this API for yourself, or if you have \'admin\' privileges');
     }
 
-    const appPermissions = await userAppPermissionService.findByUser(principal);
     ctx.response.body = hal.collection(
       principal,
-      appPermissions
+      identities,
     );
 
   }
 
 }
 
-export default new UserAppPermissionsCollection();
+export default new PrincipalIdentityCollection();