+++ src/reset-password/service.ts
@@ -1,8 +1,9 @@
 import * as nodemailer from 'nodemailer';
 import { render } from '../templates.js';
-import { User } from '../types.js';
-import { createToken } from '../one-time-token/service.js';
+import { PrincipalIdentity, User } from '../types.js';
+import { createToken } from '../verification-token/service.js';
 import { requireSetting } from '../server-settings.js';
+import { BadRequest } from '@curveball/http-errors';
 
 /**
  * This function is for sending reset password email with validated token
@@ -11,23 +12,27 @@ import { requireSetting } from '../server-settings.js';
  *
  * Renders an email with provided from, to, subject, html template.
  */
-export async function sendResetPasswordEmail(user: User) {
+export async function sendResetPasswordEmail(user: User, identity: PrincipalIdentity) {
 
   const smtpEmailFrom = requireSetting('smtp.emailFrom')!;
   const smtpUrl = requireSetting('smtp.url')!;
 
   const transporter = nodemailer.createTransport(smtpUrl);
-  const token = await createToken(user, null);
+  const token = await createToken(user, null, identity);
   const emailTemplate = render('emails/reset-password-email', {
     name: user.nickname,
     url: process.env.PUBLIC_URI + 'reset-password/token/' + token.token,
     expiryHours: token.ttl / 60 / 60
   }, false);
 
+  if (!identity.href.startsWith('mailto:')) {
+    throw new BadRequest('You can only request a password reset with an email address.');
+  }
+
   // send mail with defined transport object
   const info = await transporter.sendMail({
     from: smtpEmailFrom, // sender address
-    to: user.identity.substring(7), // list of receivers
+    to: identity.href.substring(7), // list of receivers
     subject: 'Password reset request', // Subject line
     html: emailTemplate
   });