+++ src/services.ts
@@ -1,0 +1,6 @@
+export * as principalIdentity from './principal-identity/service.js';
+export * as principal from './principal/service.js';
+export * as user from './user/service.js';
+export * as privilege from './privilege/service.js';
+export * as resetPassword from './reset-password/service.js';
+export * as log from './log/service.js';